+++ src/js/components/sortable.js
@@ -1,6 +1,6 @@
 import Animate from '../mixin/animate';
 import Class from '../mixin/class';
-import {$$, addClass, after, assign, append, attr, before, closest, css, height, getPos, includes, index, isInput, offset, off, on, pointerDown, pointerMove, pointerUp, preventClick, remove, removeClass, scrollTop, toggleClass, toNodes, trigger, within} from 'uikit-util';
+import {$$, addClass, after, assign, append, attr, before, css, height, getPos, includes, index, isInput, offset, off, on, pointerDown, pointerMove, pointerUp, preventClick, remove, removeClass, scrollTop, toggleClass, toNodes, trigger, within} from 'uikit-util';
 
 export default {
 
@@ -49,7 +49,9 @@ export default {
 
     events: {
 
-        [pointerDown]: 'init'
+        name: pointerDown,
+        passive: false,
+        handler: 'init'
 
     },
 
@@ -92,7 +94,7 @@ export default {
             const [placeholder] = toNodes(this.$el.children).filter(el => within(target, el));
 
             if (!placeholder
-                || isInput(e.target)
+                || isInput(target)
                 || this.handle && !within(target, this.handle)
                 || button > 0
                 || within(target, `.${this.clsNoDrag}`)
@@ -189,9 +191,8 @@ export default {
             off(window, 'scroll', this.scroll);
 
             if (!this.drag) {
-
-                if (e.type !== 'mouseup' && within(e.target, 'a[href]')) {
-                    location.href = closest(e.target, 'a[href]').href;
+                if (e.type === 'touchend') {
+                    e.target.click();
                 }
 
                 return;