+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.0.0 rc 23 (November 16, 2018)
+
+### Fixed
+
+- Fix sticky not recalculating after update
+- Fix clicking on items not triggering `click` event in Sortable component
+- Fix Offcanvas changing viewport upon opening
+
 ## 3.0.0 rc 22 (November 9, 2018)
 
 ### Fixed